requirements.txt
google-cloud-speech
googletrans
streamlit
soundcard
comtypes